+++ Sources/RoktUXHelper/UI/Components/ImageCarouselIndicator.swift
@@ -105,10 +105,10 @@ struct ImageCarouselIndicator: View {
     func createContainer() -> some View {
         HStack(alignment: rowPerpendicularAxisAlignment(alignItems: containerStyle?.alignItems),
                spacing: CGFloat(containerStyle?.gap ?? 0)) {
-            ForEach(model.rowViewModels) {
+            ForEach(model.rowViewModels) { viewModel in
                 RowComponent(
                     config: config,
-                    model: $0,
+                    model: viewModel,
                     parentWidth: $model.availableWidth,
                     parentHeight: $model.availableHeight,
                     styleState: $model.styleState,
@@ -116,7 +116,7 @@ struct ImageCarouselIndicator: View {
                 )
             }
         }
-               .accessibilityElement(children: .ignore)
-               .accessibilityHidden(true)
+        .accessibilityElement(children: .ignore)
+        .accessibilityHidden(true)
     }
 }